cleanup KeyMixin
And move the logic into the secret and lease models which used it
