+++ ui/app/models/secret-v2.js
@@ -7,7 +7,7 @@ import Model, { belongsTo, hasMany, attr } from '@ember-data/model';
 import { computed } from '@ember/object'; // eslint-disable-line
 import { alias } from '@ember/object/computed'; // eslint-disable-line
 import { expandAttributeMeta } from 'vault/utils/field-to-attrs';
-import KeyMixin from 'vault/mixins/key-mixin';
+import utils from 'vault/lib/key-utils';
 import lazyCapabilities, { apiPath } from 'vault/macros/lazy-capabilities';
 import { withModelValidations } from 'vault/decorators/model-validations';
 
@@ -20,7 +20,7 @@ const validations = {
 
 @withModelValidations(validations)
 class SecretV2Model extends Model {}
-export default SecretV2Model.extend(KeyMixin, {
+export default SecretV2Model.extend({
   failedServerRead: attr('boolean'),
   engine: belongsTo('secret-engine', { async: false }),
   engineId: attr('string'),
@@ -66,4 +66,53 @@ export default SecretV2Model.extend(KeyMixin, {
   canReadSecretData: alias('secretDataPath.canRead'),
   canEditSecretData: alias('secretDataPath.canUpdate'),
   canDeleteSecretData: alias('secretDataPath.canDelete'),
+
+  // what attribute has the path for the key
+  // will.be 'path' for v2 or 'id' v1
+  pathAttr: 'path',
+  flags: null,
+
+  initialParentKey: null,
+
+  isCreating: computed('initialParentKey', function () {
+    return this.initialParentKey != null;
+  }),
+
+  pathVal() {
+    return this[this.pathAttr] || this.id;
+  },
+
+  // rather than using defineProperty for all of these,
+  // we're just going to hardcode the known keys for the path ('id' and 'path')
+  isFolder: computed('id', 'path', function () {
+    return utils.keyIsFolder(this.pathVal());
+  }),
+
+  keyParts: computed('id', 'path', function () {
+    return utils.keyPartsForKey(this.pathVal());
+  }),
+
+  parentKey: computed('id', 'path', 'isCreating', {
+    get: function () {
+      return this.isCreating ? this.initialParentKey : utils.parentKeyForKey(this.pathVal());
+    },
+    set: function (_, value) {
+      return value;
+    },
+  }),
+
+  keyWithoutParent: computed('id', 'path', 'parentKey', {
+    get: function () {
+      var key = this.pathVal();
+      return key ? key.replace(this.parentKey, '') : null;
+    },
+    set: function (_, value) {
+      if (value && value.trim()) {
+        this.set(this.pathAttr, this.parentKey + value);
+      } else {
+        this.set(this.pathAttr, null);
+      }
+      return value;
+    },
+  }),
 });
